Split durable session service and prompt modules
