+++ anyharness/crates/anyharness-lib/src/sessions/service/creation.rs
@@ -1,0 +1,163 @@
+use std::time::Instant;
+
+use uuid::Uuid;
+
+use crate::domains::agents::registry::built_in_registry;
+use crate::domains::agents::resolver::resolve_agent;
+use crate::origin::OriginContext;
+use crate::sessions::model::{SessionMcpBindingPolicy, SessionRecord};
+
+use super::model_resolution::{find_model_registry, resolve_model_id};
+use super::{CreateSessionError, SessionService};
+
+impl SessionService {
+    pub fn create_session(
+        &self,
+        workspace_id: &str,
+        agent_kind: &str,
+        model_id: Option<&str>,
+        mode_id: Option<&str>,
+        mcp_bindings_ciphertext: Option<String>,
+        mcp_binding_summaries_json: Option<String>,
+        mcp_binding_policy: SessionMcpBindingPolicy,
+        system_prompt_append: Option<String>,
+        subagents_enabled: bool,
+        origin: OriginContext,
+    ) -> Result<SessionRecord, CreateSessionError> {
+        let started = Instant::now();
+        tracing::info!(
+            workspace_id = %workspace_id,
+            agent_kind = %agent_kind,
+            model_id = ?model_id,
+            mode_id = ?mode_id,
+            "[workspace-latency] session.create.validate.start"
+        );
+
+        let workspace_lookup_started = Instant::now();
+        let workspace = self
+            .workspace_store
+            .find_by_id(workspace_id)
+            .map_err(CreateSessionError::Internal)?
+            .ok_or_else(|| CreateSessionError::WorkspaceNotFound(workspace_id.to_string()))?;
+        tracing::info!(
+            workspace_id = %workspace_id,
+            elapsed_ms = workspace_lookup_started.elapsed().as_millis(),
+            "[workspace-latency] session.create.workspace_validated"
+        );
+
+        if workspace.surface == "cowork" {
+            if let Some(existing) = self
+                .session_store
+                .list_with_dismissed_by_workspace(workspace_id)
+                .map_err(CreateSessionError::Internal)?
+                .into_iter()
+                .next()
+            {
+                return Err(CreateSessionError::WorkspaceSingleSession {
+                    workspace_id: workspace_id.to_string(),
+                    session_id: existing.id,
+                });
+            }
+        }
+
+        let registry_lookup_started = Instant::now();
+        let registry = built_in_registry();
+        let descriptor = registry
+            .iter()
+            .find(|d| d.kind.as_str() == agent_kind)
+            .ok_or_else(|| {
+                CreateSessionError::Invalid(format!("unknown agent kind: {agent_kind}"))
+            })?;
+        tracing::info!(
+            workspace_id = %workspace_id,
+            agent_kind = %agent_kind,
+            elapsed_ms = registry_lookup_started.elapsed().as_millis(),
+            "[workspace-latency] session.create.agent_descriptor_found"
+        );
+
+        let agent_resolution_started = Instant::now();
+        let resolved = resolve_agent(descriptor, &self.runtime_home);
+        if resolved.status != crate::domains::agents::model::ResolvedAgentStatus::Ready {
+            let detail = resolved.agent_process.message.clone().or_else(|| {
+                resolved
+                    .native
+                    .as_ref()
+                    .and_then(|artifact| artifact.message.clone())
+            });
+            if let Some(detail) = detail {
+                return Err(CreateSessionError::Invalid(format!(
+                    "agent '{agent_kind}' is not ready (status: {:?}): {detail}",
+                    resolved.status
+                )));
+            }
+            return Err(CreateSessionError::Invalid(format!(
+                "agent '{agent_kind}' is not ready (status: {:?})",
+                resolved.status
+            )));
+        }
+        tracing::info!(
+            workspace_id = %workspace_id,
+            agent_kind = %agent_kind,
+            elapsed_ms = agent_resolution_started.elapsed().as_millis(),
+            "[workspace-latency] session.create.agent_validated"
+        );
+
+        let model_resolution_started = Instant::now();
+        let registries = self.model_catalog_service.registries();
+        let model_registry = find_model_registry(&registries, agent_kind)
+            .map_err(|error| CreateSessionError::Invalid(error.to_string()))?;
+        let resolved_model_id = resolve_model_id(model_registry, model_id)
+            .map_err(|error| CreateSessionError::Invalid(error.to_string()))?;
+        let resolved_mode_id = mode_id
+            .map(str::trim)
+            .filter(|value| !value.is_empty())
+            .map(str::to_string);
+        tracing::info!(
+            workspace_id = %workspace_id,
+            agent_kind = %agent_kind,
+            resolved_model_id = ?resolved_model_id,
+            resolved_mode_id = ?resolved_mode_id,
+            elapsed_ms = model_resolution_started.elapsed().as_millis(),
+            "[workspace-latency] session.create.model_resolved"
+        );
+
+        let now = chrono::Utc::now().to_rfc3339();
+        let record = SessionRecord {
+            id: Uuid::new_v4().to_string(),
+            workspace_id: workspace_id.to_string(),
+            agent_kind: agent_kind.to_string(),
+            native_session_id: None,
+            requested_model_id: resolved_model_id.clone(),
+            current_model_id: resolved_model_id,
+            requested_mode_id: resolved_mode_id.clone(),
+            current_mode_id: resolved_mode_id,
+            title: None,
+            thinking_level_id: None,
+            thinking_budget_tokens: None,
+            status: "starting".into(),
+            created_at: now.clone(),
+            updated_at: now,
+            last_prompt_at: None,
+            closed_at: None,
+            dismissed_at: None,
+            mcp_bindings_ciphertext,
+            mcp_binding_summaries_json,
+            mcp_binding_policy,
+            system_prompt_append,
+            subagents_enabled,
+            action_capabilities_json: None,
+            origin: Some(origin),
+        };
+
+        self.session_store
+            .insert(&record)
+            .map_err(CreateSessionError::Internal)?;
+        tracing::info!(
+            workspace_id = %workspace_id,
+            session_id = %record.id,
+            elapsed_ms = started.elapsed().as_millis(),
+            "[workspace-latency] session.create.durable_record_inserted"
+        );
+        Ok(record)
+    }
+}
